Fixes for warning: format '%d' expects type 'int', but argument 5 has type 'size_t'
git-svn-id: svn://scm.orgis.org/mpg123/trunk@548 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/id3.c
@@ -92,7 +92,7 @@ void store_id3_text(struct stringbuf* sb, char* source, size_t source_size)
 	if((source_size-1) % bwidth)
 	{
 		/* Uh. (BTW, the -1 is for the encoding byte.) */
-		warning2("Weird tag size %d for encoding %d - I will probably trim too early or something but I think the MP3 is broken.", source_size, encoding);
+		warning2("Weird tag size %d for encoding %d - I will probably trim too early or something but I think the MP3 is broken.", (int)source_size, encoding);
 		source_size -= (source_size-1) % bwidth;
 	}
 	/*
